Add utils why it is simple structure
