+++ src/.gitignore
@@ -4,7 +4,7 @@ application/cache/*
 !application/cache/index.html
 !application/cache/.htaccess
 
-application/controllers/apiKey.php
+application/controllers/api_key.php
 
 application/logs/*
 !application/logs/index.html
