+++ src/application/controllers/utils/send_message.php
@@ -1,0 +1,25 @@
+<?php
+//メッセージの送信
+function sending_messages($accessToken, $replyToken, $message_type, $return_message_text){
+    $response_format_text = [
+        "type" => $message_type,
+        "text" => $return_message_text
+    ];
+
+    $post_data = [
+        "replyToken" => $replyToken,
+        "messages" => [$response_format_text]
+    ];
+
+    $ch = curl_init("https://api.line.me/v2/bot/message/reply");
+    curl_setopt($ch, CURLOPT_POST, true);
+    curl_setopt($ch, CURLOPT_CUSTOMREQUEST, 'POST');
+    curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
+    curl_setopt($ch, CURLOPT_POSTFIELDS, json_encode($post_data));
+    curl_setopt($ch, CURLOPT_HTTPHEADER, array(
+        'Content-Type: application/json; charser=UTF-8',
+        'Authorization: Bearer ' . $accessToken
+    ));
+    $result = curl_exec($ch);
+    curl_close($ch);
+}
